+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="0.2.8"></a>
+# [0.2.8](https://github.com/raulsebastianmihaila/crizmas-router/compare/v0.2.7...v0.2.8) (2017-05-28)
+
+### Fixes
+- Fix reusing route fragments when the abstract path is empty.
+
+### Updates
+- Decode path params.
+- Small refactoring.
+
 <a name="0.2.7"></a>
 # [0.2.7](https://github.com/raulsebastianmihaila/crizmas-router/compare/v0.2.6...v0.2.7) (2017-05-14)